backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4.0
django-cors-headers>=4.7,<5.0
Pillow>=11.0,<12.0
psycopg[binary]>=3.2,<4.0
argon2-cffi>=23.1,<26.0
